#5698, #5711, #5708, #5710, #5699, #5692: Workflow Changes - Sidharth ===========================
1. Moved the placeAtCenter function from homestuff.js to general.js

2. There were some script tags were pointing to already js files, in one case th
e file paths were wrong for functional.js. These have been cleaned up.

3. Now related fields of type User (eg. assigned_to) show up in the dropdowns fo
r email task and the workflow conditions pages.

4. Workflows can be saved as templates which can be used to create new copies of
 these workflows including any tasks attached to them. There is a save template
button on the edit workflow page which will save the current workflow as a templ
ate. While creating a workflow the popup shows a radio button to select to creat
e a workflow from a template.

5. There was an issue with workflow where utf-8 characters were not getting disp
layed properly.


6. A new task has been added to create an event when a workflow is triggered eve
nts.

7. Trac ticket 5710
A cancel button has been added for tasks, this will take the user back to the wo
rkflow.


8. Trac ticket 5708
When a module is selected in the Module list view's filter select drop-down, and
 the user tries to create a new workflow the filtered module will automatically
get selected in the new module popup's module drop-down.


9. The ability to validate forms has been added, currently only restricted to ce
rtain mandatory fields. For example title in the Edit Workflow and Edit Task pag
e.


10. New tasks will be active by default.


11. Trac Ticket 5711
The time fields in Create Todo Task and Create Event Task have been made into dr
op-downs instead of input boxes.


12. Trac Ticket 5698
The edit task would display the module id along with the module title, the modul
e id has been removed.


13. Trac Ticket 5695
The user will be taken to an error page if he tries to access a non existent tas
k or workflow.


14. Only the admin user can access the pages of the workflow module.

+++ modules/com_vtiger_workflow/deletetask.php
@@ -4,9 +4,19 @@ require_once("include/events/SqlResultIterator.inc");
 require_once("include/Zend/Json.php");
 require_once("VTWorkflowApplication.inc");
 require_once("VTTaskManager.inc");
+require_once('VTWorkflowUtils.php');
 
 	function vtDeleteWorkflow($adb, $request){
-		$module = new VTWorkflowApplication("saveworkflow");
+		$util = new VTWorkflowUtils();
+		$module = new VTWorkflowApplication("deltetask");
+		$mod = return_module_language($current_language, $module->name);
+
+		if(!$util->checkAdminAccess()){
+			$errorUrl = $module->errorPageUrl($mod['LBL_ERROR_NOT_ADMIN']);
+			$util->redirectTo($errorUrl, $mod['LBL_ERROR_NOT_ADMIN']);
+			return;
+		}
+
 		$wm = new VTTaskManager($adb);
 		$wm->deleteTask($request['task_id']);
